1. move FileUtils to node-commons. 2. modify the arch of snapshot process.
