+++ node-commons/src/main/java/org/apache/iotdb/commons/auth/authorizer/IAuthorizer.java
@@ -16,18 +16,20 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+
 package org.apache.iotdb.commons.auth.authorizer;
 
 import org.apache.iotdb.commons.auth.AuthException;
 import org.apache.iotdb.commons.auth.entity.Role;
 import org.apache.iotdb.commons.auth.entity.User;
+import org.apache.iotdb.commons.snapshot.SnapshotProcessor;
 
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
 /** This interface provides all authorization-relative operations. */
-public interface IAuthorizer {
+public interface IAuthorizer extends SnapshotProcessor {
 
   /**
    * Login for a user.
