+++ node-commons/src/main/java/org/apache/iotdb/commons/auth/user/LocalFileUserAccessor.java
@@ -22,8 +22,10 @@ import org.apache.iotdb.commons.auth.entity.PathPrivilege;
 import org.apache.iotdb.commons.auth.entity.User;
 import org.apache.iotdb.commons.conf.IoTDBConstant;
 import org.apache.iotdb.commons.file.SystemFileFactory;
+import org.apache.iotdb.commons.utils.FileUtils;
 import org.apache.iotdb.commons.utils.IOUtils;
 
+import org.apache.thrift.TException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -41,6 +43,7 @@ import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import java.util.UUID;
 
 /**
  * This class loads a user's information from the corresponding file.The user file is a sequential
@@ -55,10 +58,10 @@ import java.util.Set;
  * bytes
  */
 public class LocalFileUserAccessor implements IUserAccessor {
-
+  private static final Logger logger = LoggerFactory.getLogger(LocalFileUserAccessor.class);
   private static final String TEMP_SUFFIX = ".temp";
   private static final String STRING_ENCODING = "utf-8";
-  private static final Logger logger = LoggerFactory.getLogger(LocalFileUserAccessor.class);
+  private static final String userSnapshotFileName = "system" + File.separator + "users";
 
   private String userDirPath;
   /**
@@ -232,6 +235,45 @@ public class LocalFileUserAccessor implements IUserAccessor {
       retList.addAll(set);
     }
     return retList;
+  }
+
+  @Override
+  public boolean processTakeSnapshot(File snapshotDir) throws TException, IOException {
+    SystemFileFactory systemFileFactory = SystemFileFactory.INSTANCE;
+    File userFolder = systemFileFactory.getFile(userDirPath);
+    File userSnapshotDir = systemFileFactory.getFile(snapshotDir, userSnapshotFileName);
+    File userTmpSnapshotDir =
+        systemFileFactory.getFile(userSnapshotDir.getAbsolutePath() + "-" + UUID.randomUUID());
+
+    boolean result = true;
+    try {
+      result = FileUtils.copyDir(userFolder, userTmpSnapshotDir);
+      result &= userTmpSnapshotDir.renameTo(userSnapshotDir);
+    } finally {
+      userTmpSnapshotDir.delete();
+    }
+    return result;
+  }
+
+  @Override
+  public void processLoadSnapshot(File snapshotDir) throws TException, IOException {
+    SystemFileFactory systemFileFactory = SystemFileFactory.INSTANCE;
+    File userFolder = systemFileFactory.getFile(userDirPath);
+    File userTmpFolder =
+        systemFileFactory.getFile(userFolder.getAbsolutePath() + "-" + UUID.randomUUID());
+    File userSnapshotDir = systemFileFactory.getFile(snapshotDir, userSnapshotFileName);
+
+    try {
+      userFolder.renameTo(userTmpFolder);
+      if (!FileUtils.copyDir(userSnapshotDir, userFolder)) {
+        logger.error("Failed to load user folder snapshot and rollback.");
+        // rollback if failed to copy
+        FileUtils.deleteDirectory(userFolder);
+        userTmpFolder.renameTo(userFolder);
+      }
+    } finally {
+      FileUtils.deleteDirectory(userTmpFolder);
+    }
   }
 
   @Override
