+++ server/src/main/java/org/apache/iotdb/db/engine/storagegroup/DataRegion.java
@@ -1571,7 +1571,7 @@ public class DataRegion {
     try {
       File storageGroupFolder = SystemFileFactory.INSTANCE.getFile(systemDir, dataRegionId);
       if (storageGroupFolder.exists()) {
-        org.apache.iotdb.db.utils.FileUtils.deleteDirectory(storageGroupFolder);
+        org.apache.iotdb.commons.utils.FileUtils.deleteDirectory(storageGroupFolder);
       }
     } finally {
       writeUnlock();
@@ -1628,7 +1628,7 @@ public class DataRegion {
       File storageGroupFolder =
           fsFactory.getFile(tsfilePath, logicalStorageGroupName + File.separator + dataRegionId);
       if (storageGroupFolder.exists()) {
-        org.apache.iotdb.db.utils.FileUtils.deleteDirectory(storageGroupFolder);
+        org.apache.iotdb.commons.utils.FileUtils.deleteDirectory(storageGroupFolder);
       }
     }
   }
